[Lucas][UI] Make progress bar app store downloading like. Add display label to display current page number. Remove extra code.

+++ Lucas/Lucas/RMLoader/RMDisplayLabel.h
@@ -11,6 +11,6 @@
 @interface RMDisplayLabel : UILabel
 
 // update the numbers
-- (void)updateValue:(CGFloat)value;
+- (void)updateValue:(NSInteger)value;
 
 @end
